Development (#24)
* build(frontend): fix build-fe command in package.json

* build(github workflows)

: build frontend and move it to static folder of spring

* feat(FE; github ): WIP: add gitHub action for building fe-image

* feat(github; build-docker-image ): renamed file to build-docker-images.yml and added usage of checkout@v2

* refactor(github; build-docker-images ): revert renaming to build-docker-images.yml

* feat(FE; docker-compose): update docker-compose.yml to use frontend image

* feat(FE, Docker Env Variables): Generate app-config.js at startup

* feat(FE, Docker Env Variables): Generate app-config.js at startup

* feat(FE; docker): WIP: build working fe image through Dockerfile

* feat(FE; docker): added dockerinore for faster image building

* feat(FE, Docker Env Variables): WIP Generate app-config.js at startup

* feat(FE, Docker Env Variables): WIP Working script, Missing docker entrypoint

* refactor(FE, Dockerfile): Script for env. is now started

* feat(FE, Docker Env Variables): WIP Working script, Missing docker entrypoint

* refactor(FE, Dockerfile): Script for env. is now started

* refactor(FE, Dockerfile): Script for env. is now started

* refactor(FE, Dockerfile): Script for env. is now started

* refactor(FE, nginx): Frontend now communicates with the backend properly

* refactor(FE, nginx): Frontend now communicates with the backend properly

* refactor(FE, nginx): Frontend now communicates with the backend properly

* refactor(Project, docker): Added docker-compose in project root.

* feat( Docker-compose, connect all containers ): WIP docker network

* feat( Docker-compose, connect all containers ): working docker-compose file in backend/docker folder

* feat( Docker-compose, connect all containers ): sample env files

* build(frontend): dont redirect /api in nginx.conf

* feat( Docker-compose, connect all containers ): more verbose bash script

* feat( Docker-compose, local compose setup ): WIP Docker-compose won't run migrations

* refactor(DevOp  WIP ):

* feat( Docker-compose, local compose setup ): reverted docker-compose setup

* feat(FE; index.html): Show app version only in developer tool for debug purpose

* feat(FE; index.html): Show app version only in developer tool for debug purpose

* feat(gitworkflows): added workflow for prod image building

* refactor(github workflows): refactored job names

* feature(workflow): added  workflow

* fix(MeetingController): Added check if meetingproposal is matched before deletion. Testing is WIP

* test(MeetingControllerTest.java): Added test for case if meetingProposal is not matched.

* refactor(Meetingservice, backend): Changed isproposalmatched to getById instead of findbyId

* fix(MeetingProposalValidation):Added Validation for winter and summertime

* refactor(DiversityLunchSecurityExpressionRoot): Remove debug statements from isProfileOwner

* fix(MeetingControllerIT): Fix test by updating MeetingTestdataFactory to avoid changes to the test by the introduction of summer/winter time validation constraints

* test(MeetingProposalValidation): add MeetingProposalValidationTest.java

* doc(MeetingProposalValidation): add a string documenting the functionality of  MeetingProposalValidation

* fix(MeetingProposalValidation): use Europe/Berlin time zone instead of default zone of the system to check if the date had daylight saving time

* refactor(MeetingProposalValidation): use LocalTime instead of Strings to compare times in MeetingProposalValidation

* style(MeetingProposalValidation): MeetingProposalValidation rename timeZone variable to europeBerlinTimeZone

* deleted unnecessary build-docker-image.yml

file is replaced by build-docker-dev-images.yml and build-docker-prod-images.yml

* update .github/workflows/build-docker-dev-images.yml

* fix syntax on .github/workflows/build-docker-dev-images.yml

* fix syntax for real this time on .github/workflows/build-docker-dev-images.yml

* 0.0.1: Add dev workflow in github actions, which extracts version number from the latest commit messages to build docker images.

* 0.0.1: Add prod workflow in github actions, which extracts version number from the latest commit messages to build docker images.

* refactor(PR-Changes): fulfilled Change-Requests

* refactor(PR-Changes): fulfilled Change-Requests

+++ diversity-lunch-be/src/main/java/de/brockhausag/diversitylunchspringboot/meeting/controller/MeetingController.java
@@ -5,7 +5,6 @@ import de.brockhausag.diversitylunchspringboot.meeting.model.CreateMeetingPropos
 import de.brockhausag.diversitylunchspringboot.meeting.model.MeetingDto;
 import de.brockhausag.diversitylunchspringboot.meeting.model.MeetingProposalEntity;
 import de.brockhausag.diversitylunchspringboot.meeting.service.MeetingService;
-import de.brockhausag.diversitylunchspringboot.meeting.utils.MeetingProposalValidation;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.media.Content;
 import io.swagger.v3.oas.annotations.media.Schema;
@@ -73,14 +72,18 @@ public class MeetingController {
                 .orElse(ResponseEntity.badRequest().build());
     }
 
-    @Operation(summary = "die Meetings eines registrierten Benutzers werden gelöscht")
+    @Operation(summary = "die Terminvorschläge für ein Meeting eines registrierten Benutzers werden gelöscht")
     @ApiResponses(value = {
-            @ApiResponse(responseCode = "200", description = "ein Meeting des Benutzers wurde aus der Datenbank gelöscht"),
+            @ApiResponse(responseCode = "200", description = "ein Terminvorschlag für ein Meeting wurde aus der Datenbank gelöscht"),
+            @ApiResponse(responseCode = "403", description = "für diesen Terminvorschlag existiert bereits ein aktives Meeting, deshalb kann es nicht gelöscht werden"),
     })
     @PreAuthorize("isProposalOwner(#id)")
     @DeleteMapping("/{id}")
     public @ResponseBody
     ResponseEntity<String> deleteMeetingProposal(@PathVariable Long id) {
+
+        if(this.meetingService.checkIfMeetingProposalIsMatched(id))
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN);
         this.meetingService.deleteMeetingProposal(id);
         return new ResponseEntity<>(HttpStatus.OK);
     }
